Fix Testimoni & search for berita & produks

+++ app/Http/Controllers/TestimoniController.php
@@ -4,35 +4,45 @@ namespace App\Http\Controllers;
 
 use App\Models\Testimoni;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
 
 class TestimoniController extends Controller
 {
-    public function get()
+    public function index(Request $request)
     {
-        $testimoni = Testimoni::all();
-        return view('admin.testimoni',compact('testimoni'));
+        $testimoni = Testimoni::orderBy('created_at', 'desc');
+        if ($request->has('query') && !empty($request->input('query'))) {
+            $testimoni->where('nama', 'LIKE', '%' . $request->input('query') . '%')
+                ->orWhere('asal_sekolah', 'LIKE', '%' . $request->input('query') . '%');
+        }
+
+        $testimoni = $testimoni->paginate(10);
+        return view('admin.list.testimoni.index', compact('testimoni'));
     }
 
     public function edit($id)
     {
         try {
             $test = Testimoni::FindOrFail($id);
-            return view('admin.testimonidelete',compact('test'));
+            return view('admin.list.testimoni.edit', compact('test'));
         } catch (\Throwable $th) {
-            //throw $th;
+            return back()->with('message', [
+                'icon' => 'error',
+                'title' => 'Gagal!',
+                'text' => "Ada kesalahan saat mau me edit MoU"
+            ]);
         }
     }
 
     public function create()
     {
-        return view('admin.testimonicreate');
+        return view('admin.list.testimoni.create');
     }
 
 
-    public function TestimoniStore(Request $request)
+    public function store(Request $request)
     {
-        
         $namaFoto = $request->file('foto_siswa')->hashName();
         $path = $request->file('foto_siswa')->storeAs('testimoni', $namaFoto);
         Testimoni::create([
@@ -41,18 +51,32 @@ class TestimoniController extends Controller
             'asal_sekolah' => $request->asal_sekolah,
             'testimoni' => $request->testimoni,
         ]);
-        return redirect()->back();
+        return to_route('testimoni.index')->with('message', [
+            'icon' => 'success',
+            'title' => "Berhasil!",
+            'text' => "Berhasil membuat testimoni"
+        ]);
     }
 
-    public function TestimoniUpdate(Request $request, $id)
+    public function update(Request $request, $id)
     {
         try {
-            $testimoni = Testimoni::FindOrFail($id);
+            DB::beginTransaction();
+            $testimoni = Testimoni::where('id', $id)->first();
+
+            if (!$testimoni) {
+                return back()->with('message', [
+                    'icon' => 'error',
+                    'title' => 'Gagal!',
+                    'text' => "ID testimoni tidak di temukan!"
+                ]);
+            }
+
             if ($request->hasFile('foto_siswa')) {
-                Storage::delete('testimoni/'.$testimoni->foto_siswa);
+                Storage::delete($testimoni->foto_siswa);
                 $fotoname = $request->file('foto_siswa')->hashName();
-                $path = $request->file('foto_siswa')->storeAs('testimoni',$fotoname);
-            }else{
+                $path = $request->file('foto_siswa')->storeAs('testimoni', $fotoname);
+            } else {
                 $path = $testimoni->foto_siswa;
             }
             $testimoni->nama = $request->nama;
@@ -60,21 +84,54 @@ class TestimoniController extends Controller
             $testimoni->asal_sekolah = $request->asal_sekolah;
             $testimoni->testimoni = $request->testimoni;
             $testimoni->save();
-            return redirect()->back();
+
+            DB::commit();
+            return to_route('testimoni.index')->with('message', [
+                'icon' => 'success',
+                'title' => "Berhasil!",
+                'text' => "Berhasil meupdate testimoni"
+            ]);
         } catch (\Throwable $th) {
-            return redirect()->back();
+            DB::rollBack();
+            return back()->with('message', [
+                'icon' => 'error',
+                'title' => 'Gagal!',
+                'text' => "ID testimoni tidak di temukan!"
+            ]);
         }
     }
 
-    public function deleteTestimoni($id)
+    public function destroy($id)
     {
         try {
-            $testimoni = Testimoni::FindOrFail($id);
-            Storage::delete('testimoni/'.$testimoni->foto_siswa);
+            DB::beginTransaction();
+
+            $testimoni = Testimoni::where('id',$id)->first();
+
+            if (!$testimoni) {
+                return back()->with('message', [
+                    'icon' => 'error',
+                    'title' => 'Gagal!',
+                    'text' => "ID testimoni tidak di temukan!"
+                ]);
+            }
+
+            Storage::delete($testimoni->foto_siswa);
             $testimoni->delete();
-            return redirect()->back();
+
+            DB::commit();
+            return to_route('testimoni.index')->with('message', [
+                'icon' => 'success',
+                'title' => "Berhasil!",
+                'text' => "Berhasil meupdate testimoni"
+            ]);
         } catch (\Throwable $th) {
-            return redirect()->back();
+            DB::rollBack();
+            return back()->with('message', [
+                'icon' => 'error',
+                'title' => 'Gagal!',
+                'text' => "ID testimoni tidak di temukan!"
+            ]);
         }
     }
 }
